+++ settings/lib/ui/support/cubit/support_card_state.dart
@@ -1,0 +1,7 @@
+abstract class SupportCardState {}
+
+class LoadingSupportCardState extends SupportCardState {}
+
+class EnabledSupportCardState extends SupportCardState {}
+
+class DisabledSupportCardState extends SupportCardState {}